src/input_data/__init__.py
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

"""
Input Catalog package for processing university catalog data.
"""

from course_transformer import CourseTransformer
from program_transformer import ProgramTransformer
from embedding_service import generate_embeddings_batch
from ..utils.genai import GeminiClient

__all__ = [
    'CourseTransformer',
    'ProgramTransformer',
    'GeminiClient',
    'generate_embeddings_batch',
]
