src/input_data/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
jsonify==0.5
requests==2.32.4
google-cloud-bigquery==3.23.0
google-cloud-storage==2.11.0
google-cloud-aiplatform==1.59.0
tenacity==8.2.3
